+++ src/utils/map-state.js
@@ -134,9 +134,9 @@ export default class MapState {
   pan({pos, startPos}) {
     const startPanLngLat = this._interactiveState.startPanLngLat || this._unproject(startPos);
 
-    // take the start lnglat and put it where the mouse is down.
-    assert(startPanLngLat, '`startPanLngLat` prop is required ' +
-      'for mouse pan behavior to calculate where to position the map.');
+    if (!startPanLngLat) {
+      return this;
+    }
 
     const [longitude, latitude] = this._calculateNewLngLat({startPanLngLat, pos});
 
@@ -180,13 +180,10 @@ export default class MapState {
     assert(deltaScaleY >= -1 && deltaScaleY <= 1,
       '`deltaScaleY` must be a number between [-1, 1]');
 
-    let {startBearing, startPitch} = this._interactiveState;
+    const {startBearing, startPitch} = this._interactiveState;
 
-    if (!Number.isFinite(startBearing)) {
-      startBearing = this._viewportProps.bearing;
-    }
-    if (!Number.isFinite(startPitch)) {
-      startPitch = this._viewportProps.pitch;
+    if (!Number.isFinite(startBearing) || !Number.isFinite(startPitch)) {
+      return this;
     }
 
     const {pitch, bearing} = this._calculateNewPitchAndBearing({